build/modules: remove checking for modules with full path
All in-kernel modules now use $(KCONFIG) and only out of tree modules
use the fulll path.  As such this test doesn't need to remove the path.
Since out of tree modules cannot ever be built into the kernel, the
step is unneeded for them.

Due to splitting, this couldn't be removed at the same time as adding
the alternative.  As such this is now a separate step.

Signed-off-by: Elliott Mitchell <ehem+openwrt@m5p.com>